Move OnlineUsersReq/Ack to monitor namespace
