+++ PPNetClient/Delegates.cs
@@ -48,7 +48,7 @@ namespace PPNetClient
         public delegate void ServerMonitorAck(PPNetLib.Contracts.ServerMonitorAck ack);
         public event ServerMonitorAck OnServerMonitorAck;
 
-        public delegate void OnlineUsersAck(PPNetLib.Contracts.SynchroniseAcks.OnlineUsersAck ack);
+        public delegate void OnlineUsersAck(PPNetLib.Contracts.Monitor.OnlineUsersAck ack);
         public event OnlineUsersAck OnOnlineUsersAck;
     }
 }
